move tests to test_schemas.py
